Fix after-install script
